Adding Travis build. #69
Adding Travis build.

+++ build.sh
@@ -4,12 +4,19 @@ echo "current build directory:"
 pwd
 cd ../
 
-echo "download odm distribution..."
-AUTH_TOKEN=$(curl -i -H "X-Auth-User: $AUTH_USER" -H "X-Auth-Key:$AUTH_KEY" https://dal05.objectstorage.softlayer.net/auth/v1.0 | grep -E "X-Auth-Token:" | awk {'print $2'})
-curl -O -H "X-Auth-Token: $AUTH_TOKEN" $ODM_URL
+if [ ! -f $HOME/.cache/$ODM_FILE_NAME ]; then
+    echo "ODM distribution: Starting download..."
+    AUTH_TOKEN=$(curl -i -H "X-Auth-User: $AUTH_USER" -H "X-Auth-Key:$AUTH_KEY" https://dal05.objectstorage.softlayer.net/auth/v1.0 | grep -E "X-Auth-Token:" | awk {'print $2'})
+    curl -O -H "X-Auth-Token: $AUTH_TOKEN" $ODM_URL 
+    mv $ODM_FILE_NAME $HOME/.cache/
+    echo "ODM distribution: download finished..."
+else
+    echo "ODM distribution: Loading from cache..."
+    echo "ODM distribution: Loading finished..."
+fi
 
-echo "unzip odm distribution..."
-unzip -q $ODM_FILE_NAME
+echo "unzipp odm distribution..."
+unzip -q $HOME/.cache/$ODM_FILE_NAME -d .
 
 echo "copy odm-ondocker into ODM distribution..."
 cp -R odm-ondocker install
